# .migration-backup/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AgencyOS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: health
          schema:
            type: string
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientSummary"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /clients/{id}/contracts:
    get:
      operationId: getClientContracts
      tags: [clients]
      summary: Get content contracts for a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentContract"
    post:
      operationId: createClientContract
      tags: [clients]
      summary: Create content contract for a client
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentContractInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentContract"

  # ─── Leads / Sales Funnel ─────────────────────────────────────
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads (sales pipeline)
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead (incl. stage change)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /leaves:
    get:
      operationId: listLeaveRequests
      tags: [leaves]
      summary: List leave requests
      parameters:
        - in: query
          name: userId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Leave requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leaves]
      summary: Submit a leave request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/approve:
    post:
      operationId: approveLeaveRequest
      tags: [leaves]
      summary: Approve a leave request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/reject:
    post:
      operationId: rejectLeaveRequest
      tags: [leaves]
      summary: Reject a leave request
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"

  # ─── Proposals ────────────────────────────────────────────────
...
components:
  schemas:
...
    ClientSummary:
      type: object
      required: [id, companyName, category, health, projectCount, createdAt]
      properties:
        id:
          type: string
        companyName:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        category:
          type: string
        health:
          type: string
        source:
          type: ["string", "null"]
        projectCount:
          type: integer
        createdAt:
          type: string
    Client:
      type: object
      required: [id, companyName, category, health, createdAt]
      properties:
        id:
          type: string
        companyName:
          type: string
        contactPerson:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        category:
          type: string
        health:
          type: string
        source:
          type: ["string", "null"]
        internalNotes:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        facebook:
          type: ["string", "null"]
        youtube:
          type: ["string", "null"]
        linkedin:
          type: ["string", "null"]
        brandColors:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ClientInput:
      type: object
      required: [companyName]
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        billingAddress:
          type: string
        gstin:
          type: string
        category:
          type: string
        health:
          type: string
        source:
          type: string
        internalNotes:
          type: string
        website:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        youtube:
          type: string
        linkedin:
          type: string
        brandColors:
          type: string
    ClientUpdate:
      type: object
      properties:
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        billingAddress:
          type: string
        gstin:
          type: string
        category:
          type: string
        health:
          type: string
        source:
          type: string
        internalNotes:
          type: string
        website:
          type: string
        instagram:
          type: string
        facebook:
          type: string
        youtube:
          type: string
        linkedin:
          type: string
        brandColors:
          type: string
    ContentContract:
      type: object
      required: [id, clientId, platform, postsPerMonth, reelsPerMonth, storiesPerMonth, carouselsPerMonth, contractMonths, startDate, createdAt]
      properties:
        id:
          type: string
        clientId:
          type: string
        platform:
          type: string
        postsPerMonth:
          type: integer
        reelsPerMonth:
          type: integer
        storiesPerMonth:
          type: integer
        carouselsPerMonth:
          type: integer
        captionWriting:
          type: boolean
        hashtagResearch:
          type: boolean
        scheduling:
          type: boolean
        contractMonths:
          type: integer
        startDate:
          type: string
        monthlyRetainer:
          type: ["number", "null"]
        createdAt:
          type: string
    ContentContractInput:
      type: object
      required: [platform, postsPerMonth, reelsPerMonth, storiesPerMonth, carouselsPerMonth, contractMonths, startDate]
      properties:
        platform:
          type: string
        postsPerMonth:
          type: integer
        reelsPerMonth:
          type: integer
        storiesPerMonth:
          type: integer
        carouselsPerMonth:
          type: integer
        captionWriting:
          type: boolean
        hashtagResearch:
          type: boolean
        scheduling:
          type: boolean
        contractMonths:
          type: integer
        startDate:
          type: string
        monthlyRetainer:
          type: number

    # ─── Lead ─────────────────────────────────────────────
    Lead:
      type: object
      required: [id, title, stage, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        companyName:
          type: ["string", "null"]
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        value:
          type: ["number", "null"]
        stage:
          type: string
        probability:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        assigneeName:
          type: ["string", "null"]
        daysInStage:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadInput:
      type: object
      required: [title, stage]
      properties:
        title:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        value:
          type: number
        stage:
          type: string
        probability:
          type: integer
        notes:
          type: string
        industry:
          type: string
        assigneeId:
          type: string
    LeadUpdate:
      type: object
      properties:
        title:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        value:
          type: number
        stage:
          type: string
        probability:
          type: integer
        notes:
          type: string
        industry:
          type: string
        assigneeId:
          type: string
...
    LeaveRequest:
      type: object
      required: [id, userId, type, startDate, endDate, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        userName:
          type: ["string", "null"]
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string
    LeaveRequestInput:
      type: object
      required: [type, startDate, endDate]
      properties:
        type:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string

    # ─── Proposal ─────────────────────────────────────────
...